# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: HireAI - AI Interview Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /applicants:
    get:
      operationId: listApplicants
      tags: [applicants]
      summary: List all applicants
      parameters:
        - name: jobId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [new, screening, interviewing, offer, rejected, hired]
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: List of applicants
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Applicant"
    post:
      operationId: createApplicant
      tags: [applicants]
      summary: Create a new applicant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateApplicantBody"
      responses:
        "201":
          description: Created applicant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
  /applicants/{id}:
    get:
      operationId: getApplicant
      tags: [applicants]
      summary: Get an applicant by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Applicant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
        "404":
          description: Not found
    put:
      operationId: updateApplicant
      tags: [applicants]
      summary: Update an applicant
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateApplicantBody"
      responses:
        "200":
          description: Updated applicant
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Applicant"
...
  /dashboard/recent-activity:
    get:
      operationId: getRecentActivity
      tags: [dashboard]
      summary: Get recent activity feed
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent activity
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ActivityItem"
...
components:
  schemas:
...
    Applicant:
      type: object
      properties:
        id:
          type: integer
        jobId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        resumeText:
          type: string
          nullable: true
        status:
          type: string
          enum: [new, screening, interviewing, offer, rejected, hired]
        aiScore:
          type: number
          nullable: true
        aiSummary:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
        appliedAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required: [id, jobId, name, email, status, appliedAt, updatedAt]
...
    CreateApplicantBody:
      type: object
      properties:
        jobId:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        resumeText:
          type: string
          nullable: true
        source:
          type: string
          nullable: true
      required: [jobId, name, email]
    UpdateApplicantBody:
      type: object
      properties:
        status:
          type: string
          enum: [new, screening, interviewing, offer, rejected, hired]
        aiScore:
          type: number
          nullable: true
        aiSummary:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        linkedinUrl:
          type: string
          nullable: true
        resumeText:
          type: string
          nullable: true
...
    ActivityItem:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [new_applicant, interview_scheduled, interview_completed, applicant_hired, applicant_rejected, job_posted]
        title:
          type: string
        description:
          type: string
        timestamp:
          type: string
          format: date-time
        metadata:
          type: object
          nullable: true
      required: [id, type, title, description, timestamp]
...