# lib/api-zod/src/generated/types/activityItem.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireAI - AI Interview Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ActivityItemMetadata } from "./activityItemMetadata";
import type { ActivityItemType } from "./activityItemType";

export interface ActivityItem {
  id: string;
  type: ActivityItemType;
  title: string;
  description: string;
  timestamp: Date;
  metadata?: ActivityItemMetadata;
}