# lib/api-zod/src/generated/types/updateApplicantBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * HireAI - AI Interview Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { UpdateApplicantBodyStatus } from "./updateApplicantBodyStatus";

export interface UpdateApplicantBody {
  status?: UpdateApplicantBodyStatus;
  aiScore?: number | null;
  aiSummary?: string | null;
  phone?: string | null;
  linkedinUrl?: string | null;
  resumeText?: string | null;
}